delete floating-panel file from the builder
