+++ packages/design-system/src/components/floating-panel.tsx
@@ -1,8 +1,3 @@
-/**
- * Implementations of the "Floating Panel" component from:
- * https://www.figma.com/file/sfCE7iLS0k25qCxiifQNLE/%F0%9F%93%9A-Webstudio-Library?node-id=4%3A2679&t=6Q0l4j0CBvXkuKYp-0
- */
-
 import {
   type ReactNode,
   type ComponentProps,
